% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMzRtFeatIdCol.R
\name{AddMzRtFeatIdCol}
\alias{AddMzRtFeatIdCol}
\title{Add column to compound data table combining m/z and retention time values
for a unique identifier}
\usage{
AddMzRtFeatIdCol(dt, mzColNum, rtColNum, placeAfterColNum = 0,
  colName = "mzRtFeatID")
}
\arguments{
\item{dt}{data.table object with columns corresponding the a compound's
mass to charge ratio (m/z) and retention time}

\item{mzColNum}{integer; the column index number corresponding to the m/z
column in \code{dt}}

\item{rtColNum}{integer; the column index number corresponding to the
retention time column in \code{dt}}

\item{placeAfterColNum}{integer; the new column will be inserted after the
column with this index. Default value is 0 putting the new column first.}

\item{colName}{character string; the name to be given to the new column.
Defaults to "mzRtFeatID"}
}
\value{
Returns the input data.table object with the column added
}
\description{
function for creating a featID column from the mz and RT
of each feature and placing it before the first injection
column
}
\examples{
TODO
}
