% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBuildData.R
\name{CreateBuildData}
\alias{CreateBuildData}
\title{Generate Model Building Table By Combining Training Data Sets}
\usage{
CreateBuildData(parentDir, buildDataKeyFileName, outputFolderPath)
}
\arguments{
\item{parentDir}{path; path to directory with the peak attribute
tables of the training data, and the key table}

\item{buildDataKeyFileName}{file name of the csv table with the peak
attribute file name in column 1 and the mz-rt of retained peaks file in
column 2 imported above}

\item{outputFolderPath}{path to directory where build data key will be
output}
}
\value{
returns a data table in the format required to build a prediction
model
}
\description{
Training data files are provided in a table of file names including
the peak attribute tables in the first column and the corresponding tables
accepted peaks' mz and RT values in the second column. This function
reads the files and generates the table used for model building.
}
\examples{
TODO

}
