% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutputDir.R
\name{CreateOutputDir}
\alias{CreateOutputDir}
\title{Create a new directory for output files}
\usage{
CreateOutputDir(parentDir, dirName)
}
\arguments{
\item{parentDir}{the directory (folder) path where the output folder
is to be created}

\item{dirName}{the name of the directory to be created}
}
\value{
returns the path of the newly created directory
}
\description{
Create a new directory for output files
}
