% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggThemeLcmsMetab.R
\name{ggThemeLcmsMetab}
\alias{ggThemeLcmsMetab}
\title{Standard ggplot2 theme for plots generated by lcmsMetab}
\usage{
ggThemeLcmsMetab()
}
\value{
returns a ggplot2 theme object
}
\description{
A simple coding time saver for developer plot aesthetic preferences.
}
