% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfPlotsSingleFile.R
\name{PdfPlotsSingleFile}
\alias{PdfPlotsSingleFile}
\title{Create PDF file from Plot List}
\usage{
PdfPlotsSingleFile(list.plotPage, orientation, folderPath, name)
}
\arguments{
\item{list.plotPage}{list of lists of ggplot plot objects}

\item{orientation}{"portrait" or "landscape" to indicate orientation of outuput
pdf}

\item{folderPath}{path to directory where new pdf file will be generated}

\item{name}{name of output PDF file. ".pdf" suffix will be added}
}
\value{
none
}
\description{
This function creates a pdf of the plots in list.plotPage
and exports them to the file path/name PdfFilePath
}
\examples{
TODO
}
