% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPredictionTable.R
\name{PeakPredictionTable}
\alias{PeakPredictionTable}
\title{Create Peak Prediction Table}
\usage{
PeakPredictionTable(dt.AllAtt, mzColNum = 1, rtColNum = 2)
}
\arguments{
\item{dt.AllAtt}{data.table object; contains columns:  row m/z
row retention time, for the data file elements select:
peak duration,
peak height,
peak area,
peak FWHM,
peak tailing factor,
peak asymmetry factor}

\item{mzColNum}{column index number of column containing the compound mass
to charge ratio. Defaults to 1.}

\item{rtColNum}{column index number of column containing retention time of
the compound. Defaults to 2.}
}
\value{
returns a data.table object of prediction parameters
}
\description{
This function creates a table of peak group statistics used to generate
models of peak quality and to predict peak quality from those models.
}
\examples{
TODO
}
