% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReviewBiClassPredictors.R
\name{ReviewBiClassPredictors}
\alias{ReviewBiClassPredictors}
\title{Generate Tables and Plots Examining Individual Predictors}
\usage{
ReviewBiClassPredictors(dt.featSummary.train, vect.predCols, classCol,
  outputDirPath)
}
\arguments{
\item{dt.featSummary.train}{}
}
\value{
!!!not sure yet
}
\description{
!! This function is in development
Creates a new folder in the output directory and generates useful tables
and plots for reviewing the predictive value of individual variables
}
\examples{
TODO
}
