% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTime.R
\name{runTime}
\alias{runTime}
\title{Get runtime of script}
\usage{
runTime(startTime)
}
\arguments{
\item{startTime}{sys.time() generated value from when timing should start}
}
\value{
The elapsed system time since startTime
}
\description{
Get the system time that has elapsed
since the system time of the startTime
argument
}
