% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleAttTable.R
\name{SingleAttTable}
\alias{SingleAttTable}
\title{From MZmine output of Multiple Attributes Create Single Attribute Table}
\usage{
SingleAttTable(fullAttTable, firstInjColNum, attNum, numAtts)
}
\arguments{
\item{fullAttTable}{data.table object; A data.table object in the format
exported from mzmine when multiple peak attributes are selected}

\item{firstInjColNum}{integer; The column index number of the first column
with injection peak attributes}

\item{attNum}{integer; the number corresponding to the position of the
desired attribute in the table. E.g. 3 would indicate the 3rd attribute
listed in the table.}

\item{numAtts}{integer; The number of attributes included in the table}
}
\value{
Returns the single attribute table
}
\description{
function that takes an MZmine exported table of multiple
peak attributes in data.table format and separates out a table of just the
specified attribute with the label columns.
}
\examples{
TODO
}
