% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateLogFile.R
\name{UpdateLogFile}
\alias{UpdateLogFile}
\title{create log text file and Update with new text}
\usage{
UpdateLogFile(logFilePath, logText)
}
\arguments{
\item{logFilePath}{file path of log file}

\item{logText}{text of log file}
}
\value{
No return - simply updates log file
}
\description{
create log text file and Update with new text
}
\examples{

}
