% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateLogText.R
\name{UpdateLogText}
\alias{UpdateLogText}
\title{Update summary text file}
\usage{
UpdateLogText(logText, newText, ...)
}
\arguments{
\item{logText}{character vector of existing text of log file}

\item{newText}{new text to add to the log file}

\item{...}{additional lines of text}
}
\value{
Returns text for the log file
}
\description{
Add text to log file and update text log file - used mostly for development
}
\examples{
summaryText <-
updateSummary("Initialize Summary File")
}
