% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrtTable.R
\name{WrtTable}
\alias{WrtTable}
\title{Write Data Table to CSV file}
\usage{
WrtTable(table, folderPath, name)
}
\arguments{
\item{table}{data table object}

\item{folderPath}{path to directory where table will be written}

\item{name}{name of file to be written - ".csv" will be appended}
}
\value{
Does not retun value. Writes csv file.
}
\description{
Write Data Table to CSV file
}
\examples{
TODO
}
