% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreCols.R
\name{ZScoreCols}
\alias{ZScoreCols}
\title{Converts data.table Columns to Z-scores by Column}
\usage{
ZScoreCols(dt, indFirstConvCol = 1, indLastConvCol = ncol(dt))
}
\arguments{
\item{dt}{data.table object}

\item{indFirstConvCol}{numeric; column index number of first column to be converted
z-scores. Defaults to 1.}

\item{indLastConvCol}{numeric; column index number of last column to be converted.
Defaults to the last column in the data table.}
}
\value{
Returns the adjusted data.table object
}
\description{
converts values to z scores by column for all columns between the column
indices indicated
}
\examples{
TODO
}
